Read data from DB & added auto restart on code change

+++ IOTPlanterWebService/sql/create.sql
@@ -2,7 +2,7 @@ CREATE table sensor_read (
      id SERIAL,
      plantId INTEGER,
      lux INTEGER,
-     temperature INTEGER,
+     temperature DECIMAL,
      humidity INTEGER,
      readTime BIGINT
 );
